refactor: 모델 관리 폴더 이름 변경 model => domain
